+++ src/main/java/domain/entity/task/ContentType.java
@@ -3,6 +3,8 @@ package domain.entity.task;
 /**
  * Тип контента
  */
-public enum Type {
-    VIDEO, AUDIO, PHOTO
+public enum ContentType {
+    VIDEO,
+    AUDIO,
+    PHOTO
 }